% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma.R
\name{bma}
\alias{bma}
\title{Fit nested regression submodels and compute Bayes factors (for Bayesian Model Averaging: BMA).}
\usage{
bma(full_model, prior_main_probs = NULL, prior_intr_condprobs = NULL)
}
\arguments{
\item{full_model}{An object representing the full model, i.e. the model that includes all possible terms.}

\item{prior_main_probs}{Optional: a named vector giving the prior inclusion probabilities for main effects. By default (if NULL) these are set to 0.5.}

\item{prior_intr_condprobs}{Optional: a named vector giving the prior inclusion conditional probabilities for interactions (given that the necessary main effects/lower order interactions are present. By default (if NULL) these are set to 0.5 if there are interactions, are to NULL if there are no interactions.}
}
\value{
An object of class "bma".
}
\description{
Fit nested regression submodels and compute Bayes factors (for Bayesian Model Averaging: BMA).
}
\details{
Fully Bayesian estimation is supported using "breg" model objects obtained using the "breg_laplace" or "breg_mcmc" functions (from this package).
Most maximum likelihood model fits (e.g. using the built in "glm" function) are supported with the BIC approximation.
See the vignette "intro-to-bma.Rmd" for an explanation of how to use this function and interpret results.
}
