% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg.R
\name{breg}
\alias{breg}
\title{Fit a Bayesian regression model using Stan.}
\usage{
breg(
  formula,
  data,
  center = FALSE,
  prior_scale = 1,
  seed = NA,
  chains = 4,
  iter = 5000,
  warmup = floor(iter/2)
)
}
\arguments{
\item{formula}{**}

\item{data}{**}

\item{center}{**}

\item{seed}{**}

\item{chains}{**}

\item{iter}{**}

\item{warmup}{**}

\item{r}{**}
}
\value{
A fitted Bayesian regression model (of class "breg").
}
\description{
Fit a Bayesian regression model using Stan.
}
\details{
ADD THESE
}
