% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_laplace.R
\name{breg_laplace}
\alias{breg_laplace}
\title{Fit a Bayesian regression model using the Laplace (multivariate normal) approximation.}
\usage{
breg_laplace(
  formula,
  data,
  family = "normal_linear",
  center = TRUE,
  prior_scale = 1,
  resample = FALSE,
  n_trials = NULL
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}

\item{family}{A string describing the response distribution and link function to be used in the model (see 'Details').}

\item{center}{Should the numeric predictor variables be mean-centered?}

\item{prior_scale}{Scale for the prior distribution on model coefficients (see 'Details').}

\item{resample}{Should posterior draws from the multivariate normal approximation to the posterior be resampled using importance sampling?}

\item{n_trials}{Number of trials per observation for binomial regression (not needed for other model families).}
}
\value{
A fitted Bayesian regression model (of classes "breg" and "breg_laplace").
}
\description{
Fit a Bayesian regression model using the Laplace (multivariate normal) approximation.
}
\details{
\subsection{More Information}{
More information is contained in the following vignettes:
\itemize{
\item model_families.Rmd: a list of the available types of model and a description of when to use each
\item basic_linear_regression.Rmd: shows the methods available for "breg" objects
\item intro-to-bma.Rmd: gives an explanation and example of how to use Bayesian model averaging (BMA) for variable estimation/selection
}
\subsection{Computational Details}{
This function uses Stan's optimization routines to obtain the Laplace (multivariate normal) approximation of the posterior distribution. It uses the same Stan files as breg_mcmc. Currently this function uses the Newton algorithm for optimization, because in my experience it more reliably finds the maximum than LBFGS, despite being slower. Once the model has been fit, parameters are sampled from the multivariate normal approximate posterior distribution (and transformed to the contstrained space when appropriate, e.g. for the residual standard deviation in linear models). These samples can be optionally resampled by importance sampling to improve the approximation. Parameter estimates are NOT based on these samples (this would introduce Monte Carlo error), but posterior predictions do use the samples.
}
\subsection{Prior Distributions}{
The residual variance (\eqn{\sigma^2}) has a non-informative Jeffereys prior.
The intercept (\eqn{\beta_0}) has a non-informative uniform prior.
The code does not directly put priors on the regression coefficients (\eqn{\beta_1, \beta_2, \ldots}). Instead it gives priors to standardized coefficients (\eqn{\delta_1, \delta_2, \ldots}):
\deqn{\delta_i \sim \text{Cauchy}(0, r) \text{ for } i = 1, 2, \ldots}
where \eqn{r} is given by the prior_scale function argument.
}

For models based on a normal likelihood (or related models, e.g. with censored data), these standardized parameters are defined as follows:
\deqn{\delta_i = \frac{\beta_i \text{SD}(x_i)}{\sigma} \text{ if } x_i \text{ is a real or integer predictor}}
\deqn{\delta_i = \frac{\beta_i}{\sigma} \text{ if } x_i \text{ is a contrast code for a factor}}
Note that the denominator for these standardized coefficients is not \eqn{\text{SD}(y)} (as in the standardized coefficients ordinarily reported), but instead \eqn{\sigma} (the residual standard deviation). The \eqn{\delta}'s are thus more closely related to Cohen's d.

For logistic, binomial, and ordered logistic regression models the standardized parameters are defined as:
\deqn{\delta_i = \frac{\beta_i \text{SD}(x_i)}{1.81} \text{ if } x_i \text{ is a real or integer predictor}}
\deqn{\delta_i = \frac{\beta_i}{1.81} \text{ if } x_i \text{ is a contrast code for a factor}}
Why divide by 1.81? The latent variable interpretation of logistic regression (and related models) states that:
\deqn{y = 1 \text{ if } u > 0}
where \eqn{u} is a latent variable with a standard logistic distribution. The standard logistic distribution has the following standard deviation: \eqn{\text{SD}(u) = \pi/\sqrt{3} \approx 1.81}. Thus the standardization above is consistent with the standardization for normal models: it is a generalization of Cohen's d to the latent scale.
}
}
