% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{bridge_sampler_optim}
\alias{bridge_sampler_optim}
\title{Use bridge sampling to compute the marginal likelihood of a model fit by "optimizing".
CURRENTLY THIS DOESN'T WORK}
\usage{
bridge_sampler_optim(draws_matrix, stanmodel, stan_data)
}
\arguments{
\item{draws_matrix}{**}

\item{stanmodel}{**}

\item{stan_data}{**}
}
\description{
Use bridge sampling to compute the marginal likelihood of a model fit by "optimizing".
CURRENTLY THIS DOESN'T WORK
}
