% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_data.R
\name{check_data}
\alias{check_data}
\title{Check data to make sure that it is suitable for the specified model type (family).}
\usage{
check_data(stan_data, family, n_trials)
}
\description{
Check data to make sure that it is suitable for the specified model type (family).
}
