% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{coef.breg}
\alias{coef.breg}
\title{Get parameter estimates (posterior means) from a "breg" object.}
\usage{
\method{coef}{breg}(obj)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}
}
\description{
Get parameter estimates (posterior means) from a "breg" object.
}
