% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_laplace_methods.R
\name{coef.breg_laplace}
\alias{coef.breg_laplace}
\title{Get parameter estimates (posterior means) from a "breg_laplace" object.}
\usage{
\method{coef}{breg_laplace}(obj)
}
\arguments{
\item{obj}{A "breg_laplace" object (fitted model).}
}
\description{
Get parameter estimates (posterior means) from a "breg_laplace" object.
}
