% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_mcmc_methods.R
\name{coef.breg_mcmc}
\alias{coef.breg_mcmc}
\title{Get parameter estimates (posterior means) from a "breg_mcmc" object.}
\usage{
\method{coef}{breg_mcmc}(obj, pars = c("Intercept", obj$coef_names))
}
\arguments{
\item{obj}{A "breg_mcmc" object (fitted model).}

\item{pars}{Parameters to select. By default these are just the model coefficients (fixed effects), plus the intercept.}
}
\description{
Get parameter estimates (posterior means) from a "breg_mcmc" object.
}
