% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg_methods.R
\name{coef.lapreg}
\alias{coef.lapreg}
\title{Get parameter estimates (posterior means) from a "lapreg" object.}
\usage{
\method{coef}{lapreg}(obj)
}
\arguments{
\item{obj}{A "lapreg" object (fitted model).}
}
\description{
Get parameter estimates (posterior means) from a "lapreg" object.
}
