% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_bma_methods.R
\name{coef.reg_bma}
\alias{coef.reg_bma}
\title{Get BMA estimates of coefficients.}
\usage{
\method{coef}{reg_bma}(obj)
}
\arguments{
\item{obj}{A "reg_bma" object.}
}
\value{
A table (data frame) with the following information:
mean: posterior mean
sd: posterior standard deviation
2.5 %: lower end of 95% posterior credible interval
97.5 %: upper end of 95% posterior credible interval
}
\description{
Get BMA estimates of coefficients.
}
\details{
All estimates (posterior mean, standard deviation, and credible intervals) are computed only using models that include the term in question. In other words, they should be interpreted as estimates of the coefficient IF it is included (i.e. is non-zero).
Posterior credible intervals are computed using a normal approximation.
}
