% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods_and_utilities.R
\name{dir_odds}
\alias{dir_odds}
\title{Compute Bayes factors, posterior odds etc. for directional hypotheses (with the point null hypothesis as the denominator).}
\usage{
dir_odds(obj)
}
\arguments{
\item{obj}{A "bma" object.}
}
\description{
Compute Bayes factors, posterior odds etc. for directional hypotheses (with the point null hypothesis as the denominator).
}
\details{
This function is designed to be used in the BMA "summary" method with "dir_odds" or "dir_probs". It is not designed to be used by the analyst (and is therefore not exported).
}
