% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{dmix}
\alias{dmix}
\title{Probability density function for a (univariate) mixture of normals.}
\usage{
dmix(x, pi, mu, sigma)
}
\arguments{
\item{x}{The value for which we want to find the probability density.}

\item{pi}{A vector of mixture weights.}

\item{mu}{A vector of means.}

\item{sigma}{A vector of standard deviations.}
}
\value{
The probability density.
}
\description{
Probability density function for a (univariate) mixture of normals.
}
