% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods_and_utilities.R
\name{emmeans_by_model}
\alias{emmeans_by_model}
\title{Get a list of emmeans objects for submodels that include all the factors in the "factors" argument, along with their re-normalized probabilities (pi).
This function is designed only for use internally (within the package), so it is not exported.}
\usage{
emmeans_by_model(obj, factors)
}
\description{
Get a list of emmeans objects for submodels that include all the factors in the "factors" argument, along with their re-normalized probabilities (pi).
This function is designed only for use internally (within the package), so it is not exported.
}
