% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_mcmc_methods.R
\name{factor_means}
\alias{factor_means}
\title{Compute means of the linear predictor (mu = b0 + b1*x + ...) at all combinations of factor levels (i.e. expected marginal means).}
\usage{
factor_means(obj, f = NULL)
}
\arguments{
\item{obj}{A "breg_mcmc" object (fitted model).}

\item{f}{The factor for which to compute estimated marginal means, or NULL to compute estimated marginal means for all factor combinations.}
}
\value{
The estimated marginal means and their standard deviations.
}
\description{
Compute means of the linear predictor (mu = b0 + b1*x + ...) at all combinations of factor levels (i.e. expected marginal means).
}
