% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_model}
\alias{fit_model}
\title{Fit a regression model with JZS priors.}
\usage{
fit_model(
  formula,
  data,
  family = gaussian(),
  r = 0.5,
  sample = FALSE,
  seed = NA,
  chains = 4,
  iter = 2000,
  warmup = floor(iter/2)
)
}
\arguments{
\item{formula}{An object of class formula or brmsformula (or one that can be coerced to that classes): A symbolic description of the model to be fitted. The details of model specification are explained in brmsformula.}

\item{data}{An object of class data.frame (or one that can be coerced to that class) containing data of all variables used in the model.}

\item{family}{A description of the response distribution and link function to be used in the model. This can be a family function, a call to a family function or a character string naming the family. Every family function has a link argument allowing to specify the link function to be applied on the response variable. If not specified, default links are used. For details of supported families see brmsfamily. By default, a linear gaussian model is applied.}

\item{r}{The scale of the multivariate Cauchy prior distribution on the regression coefficients (a single positive number).}

\item{seed}{The seed for random number generation to make results reproducible. If NA (the default), Stan will set the seed randomly.}

\item{chains}{Number of Markov chains (defaults to 4).}

\item{iter}{Number of total iterations per chain (including warmup; defaults to 2000).}

\item{warmup}{A positive integer specifying number of warmup (aka burnin) iterations. This also specifies the number of iterations used for stepsize adaptation, so warmup draws should not be used for inference. The number of warmup should not be larger than iter and the default is iter/2.}
}
\value{
A fitted brms model.
}
\description{
Fit a regression model with JZS priors.
}
\details{
This is a wrapper around the "brm" function from brms that adds a JZS prior on model parameters  REFS *.
}
