% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_submodels.R
\name{fit_submodels}
\alias{fit_submodels}
\title{Fit restricted submodels, given a full model.}
\usage{
fit_submodels(full_model, model_list)
}
\arguments{
\item{full_model}{A brmsfit object (fitted regression model) representing the full model, i.e. the model that includes all possible terms.}

\item{model_list}{A list of submodels, returned by using the function "submodels" on "formula(full_model)".}
}
\description{
Fit restricted submodels, given a full model.
}
