% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{format_prob}
\alias{format_prob}
\title{Format probabilities.}
\usage{
format_prob(tab)
}
\arguments{
\item{tab}{A data frame or matrix of probabilities to format.}
}
\description{
Format probabilities.
}
