% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_coef_names}
\alias{get_coef_names}
\title{Get the names of coefficients from a fitted model.}
\usage{
get_coef_names(obj)
}
\arguments{
\item{obj}{A fitted regression model.}
}
\value{
The names of the regression coefficients.
}
\description{
Get the names of coefficients from a fitted model.
}
