% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods.R
\name{get_stancode.breg}
\alias{get_stancode.breg}
\title{Print the Stan code.}
\usage{
\method{get_stancode}{breg}(obj)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}
}
\value{
The Stan code (as text).
}
\description{
Print the Stan code.
}
