% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_mcmc_methods.R
\name{get_stancode.breg_mcmc}
\alias{get_stancode.breg_mcmc}
\title{Print the Stan code.}
\usage{
\method{get_stancode}{breg_mcmc}(obj)
}
\arguments{
\item{obj}{A "breg_mcmc" object (fitted model).}
}
\value{
The Stan code (as text).
}
\description{
Print the Stan code.
}
