% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{get_term_names}
\alias{get_term_names}
\title{Get the names of terms from a fitted model.}
\usage{
get_term_names(obj)
}
\arguments{
\item{obj}{A fitted regression model.}
}
\value{
The names of the terms.
}
\description{
Get the names of terms from a fitted model.
}
