% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/has_needed.R
\name{has_needed}
\alias{has_needed}
\title{Test if a set of model terms (fixed effects) contains all the needed terms one order lower than an interaction.
This is only intended to be used in the function "test_marginality".}
\usage{
has_needed(terms, intr)
}
\description{
Test if a set of model terms (fixed effects) contains all the needed terms one order lower than an interaction.
This is only intended to be used in the function "test_marginality".
}
