% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inc_tests.R
\name{inc_tests}
\alias{inc_tests}
\title{Conduct term inclusion tests. This is the most important function in the "BayesianId" package.}
\usage{
inc_tests(full_model, prior_term_probs = NULL, digits_to_round = 2)
}
\arguments{
\item{full_model}{A brmsfit object (fitted regression model) representing the full model, i.e. the model that includes all possible terms.}

\item{prior_term_probs}{Optional: a named vector giving the prior inclusion probabilities for parameters (names = parameters/terms). By default (if NULL) these are set to 0.5, which corresponds to uniform prior probabilities across models DOUBLE CHECK *.}

\item{digits_to_round}{Number of digits to round results to.}
}
\value{
A list containing the following two elements: inc_table (term inclusion prior and posterior probabilities/odds, Bayes factors), model_table (model prior and posterior probabilities/odds, Bayes factors)
}
\description{
Conduct term inclusion tests. This is the most important function in the "BayesianId" package.
}
