% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{is.wholenumber}
\alias{is.wholenumber}
\title{Check if a number is an integer (not the integer data type, but actually an integer); this is based on the example in the documentation for "is.integer".}
\usage{
is.wholenumber(x, tol = .Machine$double.eps^0.5)
}
\description{
Check if a number is an integer (not the integer data type, but actually an integer); this is based on the example in the documentation for "is.integer".
}
