% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{is_interaction}
\alias{is_interaction}
\title{Test whether terms are interactions are not.}
\usage{
is_interaction(terms)
}
\arguments{
\item{terms}{A vector of model terms.}
}
\value{
A logical vector specifying whether each term is an interaction.
}
\description{
Test whether terms are interactions are not.
}
