% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg.R
\name{lapreg}
\alias{lapreg}
\title{Fit a Bayesian regression model using the Laplace (multivariate normal) approximation (with Stan's optimization functionality).}
\usage{
lapreg(formula, data, family = "normal_linear", center = TRUE, prior_scale = 1)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}

\item{family}{A string describing the response distribution and link function to be used in the model (see 'Details').}

\item{center}{Should the numeric predictor variables be mean-centered?}

\item{prior_scale}{Scale for the prior distribution on model coefficients (see 'Details').}
}
\value{
A fitted Bayesian regression model (of class "lapreg").
}
\description{
Fit a Bayesian regression model using the Laplace (multivariate normal) approximation (with Stan's optimization functionality).
}
\details{
ADD THIS
}
