% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{lse}
\alias{lse}
\title{Log-sum-exp trick (stable computation of the log of a sum of exponentials).}
\usage{
lse(a)
}
\arguments{
\item{a}{A vector of numbers.}
}
\value{
log-sum-exp of the vector a.
}
\description{
Log-sum-exp trick (stable computation of the log of a sum of exponentials).
}
\details{
The log-sum-exp function is defined as:
\deqn{\text{LSE}(\vec{a}) = \log\Big(\sum_j \exp(a_j)\Big)}
If the components of \eqn{\vec{a}} are large, then this will not be numerically stable.
To compute the log-sum-exp in a numerically stable way, we can use the equivalent expression:
\deqn{\text{LSE}(\vec{a}) = \max(\vec{a}) + \log\Big(\sum_j \exp(a_j - \max(\vec{a}))\Big)}
This is what the function below does.
}
