% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data.R
\name{make_stan_data}
\alias{make_stan_data}
\title{Set up a list of data for Stan.}
\usage{
make_stan_data(
  formula_info,
  data,
  family = "normal_linear",
  prior_scale = 1,
  n_trials = NULL
)
}
\description{
Set up a list of data for Stan.
}
