% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_X.R
\name{make_X}
\alias{make_X}
\title{Set up the model matrix ("X") in the form suitable for passing to stan_data.}
\usage{
make_X(formula, data, center)
}
\arguments{
\item{formula}{An object of class formula or brmsformula (or one that can be coerced to that classes): A symbolic description of the model to be fitted. The details of model specification are explained in brmsformula.}

\item{data}{A data frame.}

\item{center}{****}
}
\value{
The model matrix, with factors given appropriate contrasts for Bayesian inference, numeric predictors mean-centered (optionally), and without the intercept column.
}
\description{
Set up the model matrix ("X") in the form suitable for passing to stan_data.
}
