% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{model.frame.breg}
\alias{model.frame.breg}
\title{Return the data frame used for fitting the model.}
\usage{
\method{model.frame}{breg}(obj)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}
}
\value{
The model data frame.
}
\description{
Return the data frame used for fitting the model.
}
\details{
Note that this will return any variables that have been mean-centered in the model fitting process as mean-centered.
}
