% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg_methods.R
\name{model.frame.lapreg}
\alias{model.frame.lapreg}
\title{Return the data frame used for fitting the model.}
\usage{
\method{model.frame}{lapreg}(obj)
}
\arguments{
\item{obj}{A "lapreg" object (fitted model).}
}
\value{
The model data frame.
}
\description{
Return the data frame used for fitting the model.
}
\details{
Note that this will return any variables that have been mean-centered in the model fitting process as mean-centered.
}
