% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bfs.R
\name{model_bfs}
\alias{model_bfs}
\title{Compute Bayes factors for a list of models.}
\usage{
model_bfs(fit_list)
}
\arguments{
\item{model_list}{A list of fitted regression models, either brmsfit objects or MLE fits from "lm", "glm", "lmer" etc.}
}
\value{
A vector of Bayes factors.
}
\description{
Compute Bayes factors for a list of models.
}
\details{
The first model in the list is used as the denominator for comparisons, i.e. the Bayes factor for model i is defined as p(D | first model)/p(D | model i). If the fitted models are brmsfit objects, the Bayes factors are computed using bridge sampling. If not, it is assummed that they were fitted by MLE, and Bayes factors are computed using the BIC approximation.
}
