% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_bfs_sdr.R
\name{model_bfs_sdr}
\alias{model_bfs_sdr}
\title{Compute Bayes factors for submodels using the Savage-Dickey Density Ratio.}
\usage{
model_bfs_sdr(full_model, model_list)
}
\arguments{
\item{full_model}{A brmsfit object (fitted regression model) representing the full model, i.e. the model that includes all possible terms.}

\item{model_list}{A list of submodels, returned by using the function "submodels" on "formula(full_model)".}
}
\description{
Compute Bayes factors for submodels using the Savage-Dickey Density Ratio.
}
\details{
For now, this only works for fixed effects models. FINISH
}
