% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_log_evidence.R
\name{model_log_evidence}
\alias{model_log_evidence}
\title{Compute log model evidence (marginal likelihood) for a list of models.}
\usage{
model_log_evidence(fit_list)
}
\arguments{
\item{model_list}{A list of fitted regression models, either brmsfit objects or MLE fits from "lm", "glm", "lmer" etc.}
}
\value{
A vector of log Bayes factors.
}
\description{
Compute log model evidence (marginal likelihood) for a list of models.
}
\details{
If the fitted models are brmsfit objects, the log model evidence is computed using bridge sampling. If not, it is assummed that they were fitted by MLE, and log model evidence is computed using the BIC approximation.
}
