% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/par_tests.R
\name{par_tests}
\alias{par_tests}
\title{Conduct parameter inclusion tests. This is the most important function in the "BayesianId" package.}
\usage{
par_tests(brmsfit, prior_par_probs = NULL)
}
\arguments{
\item{brmsfit}{A brmsfit object (fitted regression model).}

\item{prior_par_probs}{Optional: a named vector giving the prior inclusion probabilities for parameters (names = parameters/terms). By default (if NULL) these are set to 0.5, which corresponds to uniform prior probabilities across models DOUBLE CHECK *.}

\item{bf_method}{Specifies the method used for computing Bayes factors. Current options are ** LIST AND GIVE DETAILS}
}
\value{
A list containing the following two elements: par_table (parameter prior and posterior probabilities/odds), model_table (model prior and posterior probabilities/odds, Bayes factors)
}
\description{
Conduct parameter inclusion tests. This is the most important function in the "BayesianId" package.
}
