% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods.R
\name{plot.bma}
\alias{plot.bma}
\title{Plot the posterior distribution of estimates.}
\usage{
\method{plot}{bma}(obj, term)
}
\arguments{
\item{obj}{A "bma" object.}

\item{coef}{The coefficient whose posterior distribution you want to plot.}
}
\description{
Plot the posterior distribution of estimates.
}
\details{
The posterior distribution is approximated with a mixture of normal distributions.
}
