% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{plot.breg}
\alias{plot.breg}
\title{Create a posterior predictive plot for diagnostic purposes.}
\usage{
\method{plot}{breg}(obj, group = NA, xvar = NA)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}

\item{group}{Name of a grouping variable (optional).}

\item{xvar}{Name of an x-axis variable for a scatterplot (optional).}
}
\description{
Create a posterior predictive plot for diagnostic purposes.
}
\details{
The type of plot created depends on the predicted/dependent variable ("y") and whether "group" or "x" is specified.

If "xvar" is specified, then the data is represented as a scatterplot with "xvar" on the x-axis and both "y" (the real data) and "yrep" (points from one simulated data set) on the y-axis. This plot uses "ppc_intervals" from bayesplot, but with the intervals removed (and only one data set) because I find them visually distracting.

If "xvar" is not specified and "y" is binary or categorical, then you get a bar plot showing the actual data counts along with the median count from the simulated data. This uses "ppc_bars" with the intervals removed (again, I find them visually distracting).

If "xvar" is not specified and "y" is numeric, then you get kernel density estimates of the distribution of the real data ("y", darker line) and the simulated data ("yrep", lighter lines, one for each simulated data set).

If "group" is specified then there will be a subplot for each level of the grouping variable (using the "_grouped" version of the appropriate bayesplot function).
}
