% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg_methods.R
\name{plot.lapreg}
\alias{plot.lapreg}
\title{Create a posterior predictive plot for diagnostic purposes.}
\usage{
\method{plot}{lapreg}(obj)
}
\arguments{
\item{obj}{A "lapreg" object (fitted model).}
}
\description{
Create a posterior predictive plot for diagnostic purposes.
}
