% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{pmix}
\alias{pmix}
\title{Cumulative distribution function for a (univariate) mixture of normals.}
\usage{
pmix(q, pi, mu, sigma)
}
\arguments{
\item{q}{The quantile for which we want to find the probability.}

\item{pi}{A vector of mixture weights.}

\item{mu}{A vector of means.}

\item{sigma}{A vector of standard deviations.}
}
\value{
The probability.
}
\description{
Cumulative distribution function for a (univariate) mixture of normals.
}
