% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_mcmc_methods.R
\name{posterior_interval.breg_mcmc}
\alias{posterior_interval.breg_mcmc}
\title{Compute posterior credible intervals for model parameters.}
\usage{
\method{posterior_interval}{breg_mcmc}(obj, prob = 0.9, pars = c("Intercept", obj$coef_names))
}
\arguments{
\item{obj}{A "breg_mcmc" object (fitted model).}

\item{prob}{Probability of the interval (e.g. 0.9 for a 90\% interval).}

\item{pars}{Parameters to select. By default these are just the model coefficients (fixed effects) plus the intercept.}
}
\value{
Posterior credible intervals for selected parameters.
}
\description{
Compute posterior credible intervals for model parameters.
}
