% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg_methods.R
\name{posterior_interval.lapreg}
\alias{posterior_interval.lapreg}
\title{Compute posterior credible intervals for model parameters (intercept plus fixed effects).}
\usage{
\method{posterior_interval}{lapreg}(obj, prob = 0.9)
}
\arguments{
\item{obj}{A "lapreg" object (fitted model).}

\item{prob}{Probability of the interval (e.g. 0.9 for a 90\% interval).}
}
\value{
Posterior credible intervals.
}
\description{
Compute posterior credible intervals for model parameters (intercept plus fixed effects).
}
