% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{posterior_linpred.breg}
\alias{posterior_linpred.breg}
\title{Draw from the posterior distribution of the linear predictor.}
\usage{
\method{posterior_linpred}{breg}(obj, newdata = NULL, ndraws = NULL)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}

\item{newdata}{Optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted data are used.}

\item{ndraws}{Number of posterior draws to use, i.e. number of replicated data sets to simulate. Defaults to NULL (use all).}
}
\value{
A D x N matrix of samples from the posterior distribution of the linear predictor, where D is the number of draws and N is the number of data points.
}
\description{
Draw from the posterior distribution of the linear predictor.
}
