% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{posterior_predict.breg}
\alias{posterior_predict.breg}
\title{Draw from the posterior predictive distribution.}
\usage{
\method{posterior_predict}{breg}(
  obj,
  newdata = NULL,
  ndraws = NULL,
  seed = sample.int(.Machine$integer.max, size = 1L)
)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}

\item{newdata}{Optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted data are used.}

\item{ndraws}{Number of posterior draws to use, i.e. number of replicated data sets to simulate. Defaults to NULL (use all).}

\item{seed}{The seed for random number generation. Set this manually if you want reproducible results.}
}
\value{
A D x N matrix of samples from the posterior predictive distribution, where D is the number of draws and N is the number of data points.
}
\description{
Draw from the posterior predictive distribution.
}
