% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_mcmc_methods.R
\name{predict.breg_mcmc}
\alias{predict.breg_mcmc}
\title{Predicted values (means and optionally SD's) based on a fitted model.}
\usage{
\method{predict}{breg_mcmc}(obj, newdata = NULL, compute_sd = FALSE)
}
\arguments{
\item{obj}{A "breg_mcmc" object (fitted model).}

\item{newdata}{Optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted linear predictors are used.}

\item{compute_sd}{Indicates whether predictive standard deviations should be computed.}
}
\value{
If compute_sd is FALSE, a vector of posterior predictive means. If compute_sd is TRUE, a list with the posterior predictive means and standard deviations.
}
\description{
Predicted values (means and optionally SD's) based on a fitted model.
}
\details{
This method is designed to be analogous to the "predict" method of lm and glm model objects. The "posterior_predict" method samples from the posterior predictive distribution; "predict" summarizes these samples to produce point predictions (means) and optionally their standard deviations (roughly analogous to predictive standard errors, but Bayesian).
}
