% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for regression by giving factors appropriate contrast codes and (optionally) mean-centering numeric predictors.}
\usage{
prepare_data(data, formula, center = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{formula}{An object of class formula or brmsformula (or one that can be coerced to that classes): A symbolic description of the model to be fitted. The details of model specification are explained in brmsformula.}

\item{center}{Should the numeric (non-factor) predictor variables be mean-centered? Defaults to TRUE.}
}
\value{
A list with the following elements:
data: a modified data frame that is ready for regression
x_numeric_names: names of data variables that are numeric
x_factor_names: names of data variables that are factors
}
\description{
Prepare data for regression by giving factors appropriate contrast codes and (optionally) mean-centering numeric predictors.
}
\details{
* DESCRIBE THE CONTRAST CODES ETC
}
