% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods_and_utilities.R
\name{print.bma}
\alias{print.bma}
\title{Print basic information about a "bma" object.}
\usage{
\method{print}{bma}(obj)
}
\arguments{
\item{obj}{A "bma" object.}
}
\description{
Print basic information about a "bma" object.
}
