% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods.R
\name{print.breg}
\alias{print.breg}
\title{Print basic information about a "breg" object.}
\usage{
\method{print}{breg}(obj)
}
\arguments{
\item{obj}{A "breg" object.}
}
\description{
Print basic information about a "breg" object.
}
