% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_laplace_methods.R
\name{print.breg_laplace}
\alias{print.breg_laplace}
\title{Print basic information about a "breg_laplace" object.}
\usage{
\method{print}{breg_laplace}(obj)
}
\arguments{
\item{obj}{A "breg_laplace" object (fitted model).}
}
\description{
Print basic information about a "breg_laplace" object.
}
