% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_mcmc_methods.R
\name{print.breg_mcmc}
\alias{print.breg_mcmc}
\title{Print basic information about a "breg_mcmc" object.}
\usage{
\method{print}{breg_mcmc}(obj)
}
\arguments{
\item{obj}{A "breg_mcmc" object (fitted model).}
}
\description{
Print basic information about a "breg_mcmc" object.
}
