% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg_methods.R
\name{print.lapreg}
\alias{print.lapreg}
\title{Print basic information about a "lapreg" object.}
\usage{
\method{print}{lapreg}(obj)
}
\arguments{
\item{obj}{A "lapreg" object (fitted model).}
}
\description{
Print basic information about a "lapreg" object.
}
