% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_bma_methods.R
\name{print.reg_bma}
\alias{print.reg_bma}
\title{Print basic information about a "reg_bma" object.}
\usage{
\method{print}{reg_bma}(obj)
}
\arguments{
\item{obj}{A "reg_bma" object.}
}
\description{
Print basic information about a "reg_bma" object.
}
