% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs_model_to_par.R
\name{probs_model_to_par}
\alias{probs_model_to_par}
\title{Convert model probabilities into parameter probablities.}
\usage{
probs_model_to_par(model_probs, included_table)
}
\arguments{
\item{model_probs}{A named vector of model probabilities (names = model formulas).}

\item{model_list}{A list containing information about some models (see restricted_models for details).}
}
\value{
A named vector of parameter inclusion probabilities (names = model terms/parameters).
}
\description{
Convert model probabilities into parameter probablities.
}
