% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs_term_to_model.R
\name{probs_term_to_model}
\alias{probs_term_to_model}
\title{Convert parameter probabilities into model probablities.}
\usage{
probs_term_to_model(term_probs, model_list)
}
\arguments{
\item{term_probs}{A named vector of parameter inclusion probabilities (names = model terms/parameters).}

\item{model_list}{A list containing information about some models (see restricted_models for details).}
}
\value{
A named vector of model probabilities (names = model formulas).
}
\description{
Convert parameter probabilities into model probablities.
}
\details{
This assumes that the prior probabilities for parameter inclusion are all independent of each other.
}
