% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{qmix}
\alias{qmix}
\title{Quantile function for a (univariate) mixture of normals.}
\usage{
qmix(p, pi, mu, sigma)
}
\arguments{
\item{p}{The probability for which we want to find the quantile.}

\item{pi}{A vector of mixture weights.}

\item{mu}{A vector of means.}

\item{sigma}{A vector of standard deviations.}
}
\value{
The quantile.
}
\description{
Quantile function for a (univariate) mixture of normals.
}
