% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_bma.R
\name{reg_bma}
\alias{reg_bma}
\title{Fit nested regression submodels and compute Bayes factors (for Bayesian Model Averaging: BMA).}
\usage{
reg_bma(
  full_model,
  prior_main_probs = NULL,
  prior_intr_condprobs = NULL,
  digits_to_round = 3
)
}
\arguments{
\item{full_model}{An object representing the full model, i.e. the model that includes all possible terms.}

\item{prior_main_probs}{Optional: a named vector giving the prior inclusion probabilities for main effects. By default (if NULL) these are set to 0.5.}

\item{prior_intr_condprobs}{Optional: a named vector giving the prior inclusion conditional probabilities for interactions (given that the necessary main effects/lower order interactions are present. By default (if NULL) these are set to 0.5 if there are interactions, are to NULL if there are no interactions.}
}
\value{
An object of class "reg_bma".
}
\description{
Fit nested regression submodels and compute Bayes factors (for Bayesian Model Averaging: BMA).
}
\details{
Fully Bayesian estimation is supported using brms.
Maximum likelihood estimates are supported (with the BIC approximation), so long as they have the following methods:
\itemize{
 \item update
 \item vcov
 \item coef
 \item predict
 \item loglik
 \item nobs
}
** CHECK AND ADD Additional details...
}
