% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted_model_bfs_sdr.R
\name{restricted_model_bfs_sdr}
\alias{restricted_model_bfs_sdr}
\title{Use the Savage-Dickey Ratio with a brms model to compute Bayes factors for restricted models.}
\usage{
restricted_model_bfs_sdr(brms_model)
}
\arguments{
\item{brms_model}{A fitted brms model.}
}
\value{
A table (data frame) of approximate Bayes factors.
}
\description{
Use the Savage-Dickey Ratio with a brms model to compute Bayes factors for restricted models.
}
\details{
Uses a multivariate normal approximation with medians and robust covariance estimation.
}
