% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted_models.R
\name{restricted_models}
\alias{restricted_models}
\title{Find all the restricted models given a regression model formula.}
\usage{
restricted_models(formula)
}
\arguments{
\item{formula}{An object of class formula or brmsformula (or one that can be coerced to that classes): A symbolic description of the full model, i.e. the model including all terms being considered.}
}
\value{
A list containing the following elements: formula_strings = model formulas as strings, formulas = model formulas, included = parameters (terms) included in each model, omitted = parameters omitted in each model, included_table = same info as "included" except in a data frame, n_models = number of models, n_fixed = number of fixed effects, fixed_names = names of fixed effects. The models are listed in each component of the output in the same order, with the full model given first.
}
\description{
Find all the restricted models given a regression model formula.
}
\details{
GIVE DETAILS ABOUT THE PRINCIPLE OF MARGINALITY *
}
