% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods_and_utilities.R
\name{show_emmeans}
\alias{show_emmeans}
\title{Get BMA estimates of factor level means (for interpreting factors).}
\usage{
show_emmeans(obj, factors, pretty = TRUE)
}
\arguments{
\item{obj}{A "bma" object.}

\item{factors}{The names(s) of the factor or factors for which to compute means.}

\item{pretty}{Logical. If TRUE, then the output is printed in an easy to read format, but of the "character" data type. If FALSE then the raw numeric output is returned, without rounding etc.}
}
\value{
A table (data frame) with the following information:
\itemize{
\item mean: posterior mean
\item sd: posterior standard deviation
\item 2.5 \%: lower end of 95\% posterior credible interval
\item 97.5 \%: upper end of 95\% posterior credible interval
}
}
\description{
Get BMA estimates of factor level means (for interpreting factors).
}
\details{
This is based on the emmeans package. That package is used to compute the expected marginal means and posterior standard deviations, which are then combined using Bayesian model averaging (BMA).
All estimates (posterior mean, standard deviation, and credible intervals) are computed only using models that include all of the relevant factors.
Posterior credible intervals are computed using a normal approximation.
}
