% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{show_stancode}
\alias{show_stancode}
\title{Print the Stan code.}
\usage{
show_stancode(obj)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}
}
\value{
The Stan code (as text).
}
\description{
Print the Stan code.
}
