% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{simulate.breg}
\alias{simulate.breg}
\alias{simulate}
\title{Simulate the model to get generated quantities (posterior predictive samples and posterior samples of mu, i.e. "linpred").}
\usage{
\method{simulate}{breg}(
  obj,
  newdata = NULL,
  ndraws = NULL,
  seed = sample.int(.Machine$integer.max, size = 1L)
)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}

\item{newdata}{Optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted data are used.}

\item{ndraws}{Number of posterior draws to use, i.e. number of replicated data sets to simulate. Defaults to NULL (use all).}

\item{seed}{The seed for random number generation. Set this manually if you want reproducible results.}
}
\value{
A list with the stanfit object (containing the samples) and "N_tilde" (the number of simulated data points).
}
\description{
Simulate the model to get generated quantities (posterior predictive samples and posterior samples of mu, i.e. "linpred").
}
\details{
This runs the appropriate "_sim" Stan code with the "gqs" rstan function, producing samples of "mu" (the linear predictor) and "Y_tilde" (the posterior predictive distribution). The user will not ordinarily use this function directly. Instead, this function is used for the "posterior_predict" and "posterior_linpred" methods.
}
