% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_laplace_methods.R
\name{simulate.breg_laplace}
\alias{simulate.breg_laplace}
\title{Simulate from the posterior predictive distribution.}
\usage{
\method{simulate}{breg_laplace}(obj)
}
\arguments{
\item{obj}{A "breg_laplace" object (fitted model).}
}
\description{
Simulate from the posterior predictive distribution.
}
\details{
Currently this is very limited, and is just a support for posterior predictive plots. In the future I should add the option to use new data.
}
