% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg_methods.R
\name{simulate.lapreg}
\alias{simulate.lapreg}
\title{Simulate from the posterior predictive distribution.}
\usage{
\method{simulate}{lapreg}(obj)
}
\arguments{
\item{obj}{A "lapreg" object (fitted model).}
}
\description{
Simulate from the posterior predictive distribution.
}
\details{
Currently this is very limited, and is just a support for posterior predictive plots. In the future I should add the option to use new data.
}
