% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_beta_draws.R
\name{std_beta_draws}
\alias{std_beta_draws}
\title{Get standardized regression coefficient draws from a brms model.}
\usage{
std_beta_draws(brms_model)
}
\arguments{
\item{brms_model}{A fitted brms model (with a sigma parameter).}
}
\value{
Draws of standardized betas (= beta/sigma), ignoring the intercept.
}
\description{
Get standardized regression coefficient draws from a brms model.
}
