% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods.R
\name{summary.breg}
\alias{summary.breg}
\title{Show summary information about a "breg" object.}
\usage{
\method{summary}{breg}(obj)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}
}
\description{
Show summary information about a "breg" object.
}
\details{
Note that the "se_mean" column represents Monte Carlo standard error, i.e. error due to MCMC estimation. The column "sd" represents posterior uncertainty, and is analogous to standard error in frequentist estimation.
}
