% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_laplace_methods.R
\name{summary.breg_laplace}
\alias{summary.breg_laplace}
\title{Show summary information about a "breg_laplace" object.}
\usage{
\method{summary}{breg_laplace}(obj)
}
\arguments{
\item{obj}{A "breg_laplace" object (fitted model).}
}
\description{
Show summary information about a "breg_laplace" object.
}
