% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_mcmc_methods.R
\name{summary.breg_mcmc}
\alias{summary.breg_mcmc}
\title{Show summary information about a "breg_mcmc" object.}
\usage{
\method{summary}{breg_mcmc}(obj)
}
\arguments{
\item{obj}{A "breg_mcmc" object (fitted model).}
}
\description{
Show summary information about a "breg_mcmc" object.
}
\details{
Note that the "se_mean" column represents Monte Carlo standard error, i.e. error due to MCMC estimation. The column "sd" represents posterior uncertainty, and is analogous to standard error in frequentist estimation.
}
