% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg_methods.R
\name{summary.lapreg}
\alias{summary.lapreg}
\title{Show summary information about a "lapreg" object.}
\usage{
\method{summary}{lapreg}(obj)
}
\arguments{
\item{obj}{A "lapreg" object (fitted model).}
}
\description{
Show summary information about a "lapreg" object.
}
