% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{terms.breg}
\alias{terms.breg}
\title{Get model "terms".}
\usage{
\method{terms}{breg}(obj)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}
}
\value{
Model "terms" object.
}
\description{
Get model "terms".
}
