% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_laplace_methods.R
\name{terms.breg_laplace}
\alias{terms.breg_laplace}
\title{Get model "terms".}
\usage{
\method{terms}{breg_laplace}(obj)
}
\arguments{
\item{obj}{A "breg_laplace" object (fitted model).}
}
\value{
Model "terms" object.
}
\description{
Get model "terms".
}
