% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_mcmc_methods.R
\name{terms.breg_mcmc}
\alias{terms.breg_mcmc}
\title{Get model "terms".}
\usage{
\method{terms}{breg_mcmc}(obj)
}
\arguments{
\item{obj}{A "breg_mcmc" object (fitted model).}
}
\value{
Model "terms" object.
}
\description{
Get model "terms".
}
