% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg_methods.R
\name{terms.lapreg}
\alias{terms.lapreg}
\title{Get model "terms".}
\usage{
\method{terms}{lapreg}(obj)
}
\arguments{
\item{obj}{A "lapreg" object (fitted model).}
}
\value{
Model "terms" object.
}
\description{
Get model "terms".
}
