% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_marginality.R
\name{test_marginality}
\alias{test_marginality}
\title{Test if a set of model terms obeys the principle of marginality, i.e. that every interaction has all the needed lower order interactions/main effects.
This is only intended to be used in the function "submodels".}
\usage{
test_marginality(terms)
}
\description{
Test if a set of model terms obeys the principle of marginality, i.e. that every interaction has all the needed lower order interactions/main effects.
This is only intended to be used in the function "submodels".
}
