% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{vcov.breg}
\alias{vcov.breg}
\title{Return the posterior variance-covariance matrix of model parameters (intercept plus fixed effects).}
\usage{
\method{vcov}{breg}(obj)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}
}
\value{
The posterior variance-covariance matrix.
}
\description{
Return the posterior variance-covariance matrix of model parameters (intercept plus fixed effects).
}
