% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_laplace_methods.R
\name{vcov.breg_laplace}
\alias{vcov.breg_laplace}
\title{Return the posterior variance-covariance matrix of model parameters (intercept plus fixed effects).}
\usage{
\method{vcov}{breg_laplace}(obj)
}
\arguments{
\item{obj}{A "breg_laplace" object (fitted model).}
}
\value{
The posterior variance-covariance matrix.
}
\description{
Return the posterior variance-covariance matrix of model parameters (intercept plus fixed effects).
}
