% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_mcmc_methods.R
\name{vcov.breg_mcmc}
\alias{vcov.breg_mcmc}
\title{Return the posterior variance-covariance matrix of model parameters.}
\usage{
\method{vcov}{breg_mcmc}(obj, pars = c("Intercept", obj$coef_names))
}
\arguments{
\item{obj}{A "breg_mcmc" object (fitted model).}

\item{pars}{Parameters to select. By default these are just the model coefficients (fixed effects) plus the intercept.}
}
\value{
The posterior variance-covariance matrix.
}
\description{
Return the posterior variance-covariance matrix of model parameters.
}
