% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapreg_methods.R
\name{vcov.lapreg}
\alias{vcov.lapreg}
\title{Return the posterior variance-covariance matrix of model parameters (intercept plus fixed effects).}
\usage{
\method{vcov}{lapreg}(obj)
}
\arguments{
\item{obj}{A "lapreg" object (fitted model).}
}
\value{
The posterior variance-covariance matrix.
}
\description{
Return the posterior variance-covariance matrix of model parameters (intercept plus fixed effects).
}
