% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg.R
\name{breg}
\alias{breg}
\title{Fit a Bayesian regression model using Stan.}
\usage{
breg(
  formula,
  data,
  family = "normal_linear",
  center = TRUE,
  prior_scale = 1,
  seed = NA,
  chains = 4,
  iter = 10000,
  warmup = floor(iter/4)
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}

\item{family}{A string describing the response distribution and link function to be used in the model (see 'Details').}

\item{center}{Should the numeric predictor variables be mean-centered?}

\item{prior_scale}{Scale for the prior distribution on model coefficients (see 'Details').}

\item{seed}{Random seed (set manually for reproducible results).}

\item{chains}{Number of MCMC chains (see rstan documentation).}

\item{iter}{Number of MCMC iterations (see rstan documentation).}

\item{warmup}{Number of warmup/burnin iterations (see rstan documentation).}
}
\value{
A fitted Bayesian regression model (of class "breg").
}
\description{
Fit a Bayesian regression model using Stan.
}
\details{
\subsection{Available Models}{
ADD THIS
}
\subsection{The Fitted Model Object}{
ADD THIS
}
\subsection{Prior Distributions}{
The residual variance (\eqn{\sigma^2}) has a non-informative Jefferys prior.
Priors are not put on the intercept (\eqn{\beta_0}) or regression coefficients (\eqn{\beta_1, \beta_2, \ldots}) directly. Instead they are put on a standardized version of the intercept (\eqn{\delta_0}) and standardized coefficients (\eqn{\delta_1, \delta_2, \ldots}).

For models based on a normal likelihood (or related models, e.g. with censored data), these standardized parameters are defined as follows:
\deqn{\delta_0 = \frac{\beta_0 - \bar{y}}{\sigma}}
\deqn{\delta_i = \frac{\beta_i \text{SD}(x_i)}{\sigma} \text{ if } x_i \text{ is a real or integer predictor}}
\deqn{\delta_i = \frac{\beta_i}{\sigma} \text{ if } x_i \text{ is a contrast code for a factor}}
Note that the denominator for these standardized coefficients is not \eqn{\text{SD}(y)} (as in the standardized coefficients ordinarily reported), but instead \eqn{\sigma} (the residual standard deviation). The \eqn{\delta}'s are thus more closely related to Cohen's d.

For logistic, binomial, and ordered logistic regression models the standardized parameters are defined as:
\deqn{\delta_0 = \frac{\beta_0}{1.81}}
\deqn{\delta_i = \frac{\beta_i \text{SD}(x_i)}{1.81} \text{ if } x_i \text{ is a real or integer predictor}}
\deqn{\delta_i = \frac{\beta_i}{1.81} \text{ if } x_i \text{ is a contrast code for a factor}}
Why divide by 1.81? The latent variable interpretation of logistic regression (and related models) states that:
\deqn{y = 1 \text{ if } u > 0}
where \eqn{u} is a latent variable with a standard logistic distribution. The standard logistic distribution has the following standard deviation: \eqn{\text{SD}(u) = \pi/\sqrt{3} \approx 1.81}. Thus the standardization above is consistent with the standardization for normal models: it is a generalization of Cohen's d to the latent scale. ** GIVE REFS
}
}
