% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg.R
\name{breg}
\alias{breg}
\title{Fit a Bayesian regression model using Stan.}
\usage{
breg(
  formula,
  data,
  family = "normal_linear",
  center = TRUE,
  prior_scale = 1,
  seed = NA,
  chains = 4,
  iter = 5000,
  warmup = floor(iter/2)
)
}
\arguments{
\item{formula}{An object of class "formula" (or one that can be coerced to that class): a symbolic description of the model to be fitted.}

\item{data}{A data frame containing the variables in the model.}

\item{family}{A string describing the response distribution and link function to be used in the model (see 'Details').}

\item{center}{Should the numeric predictor variables be mean-centered?}

\item{prior_scale}{Scale for the prior distribution on model coefficients (see 'Details').}

\item{seed}{Random seed (set manually for reproducible results).}

\item{chains}{Number of MCMC chains (see rstan documentation).}

\item{iter}{Number of MCMC iterations (see rstan documentation).}

\item{warmup}{Number of warmup/burnin iterations (see rstan documentation).}
}
\value{
A fitted Bayesian regression model (of class "breg").
}
\description{
Fit a Bayesian regression model using Stan.
}
\details{
ADD THESE
DESCRIBE THE FITTED MODEL OBJECT
DESCRIBE PRIORS
}
