% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods.R
\name{coef.breg}
\alias{coef.breg}
\title{Print parameter estimates (posterior means) from a "breg" object.}
\usage{
\method{coef}{breg}(obj, pars = c("Intercept", obj$coef_names))
}
\arguments{
\item{obj}{A "breg" object (fitted model).}

\item{pars}{Parameters to select. By default these are just the model coefficients (fixed effects), plus the intercept.}
}
\description{
Print parameter estimates (posterior means) from a "breg" object.
}
