% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods.R
\name{contrast.bma}
\alias{contrast.bma}
\title{Get BMA estimates of contrasts (for interpreting factors).}
\usage{
\method{contrast}{bma}(obj, factors, ref)
}
\arguments{
\item{obj}{A "bma" object.}

\item{factors}{The names(s) of the factor or factors for which to compute contrasts. If multiple factor names are listed, then the corresponding interaction contrasts are computed. See the examples below.}

\item{ref}{description}
}
\value{
A table (data frame) with the following information:
mean: posterior mean
sd: posterior standard deviation
2.5 %: lower end of 95% posterior credible interval
97.5 %: upper end of 95% posterior credible interval
}
\description{
Get BMA estimates of contrasts (for interpreting factors).
}
\details{
This is based on the 'contrast' method of the emmeans package. That package is used to compute the expected marginal means and posterior standard deviations, which are then combined using Bayesian model averaging (BMA). The contrasts computed are based on the 'trt.vs.ctrl' method.
All estimates (posterior mean, standard deviation, and credible intervals) are computed only using models that include all of the relevant factors. In other words, they should be interpreted as estimates of the contrasts IF they are all non-zero.
Posterior credible intervals are computed using a normal approximation.
** ADD EXAMPLES
}
