% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods_and_utilities.R
\name{dir_tests}
\alias{dir_tests}
\title{Compute Bayes factors, posterior odds etc. (or posterior/prior probabilities) for directional hypotheses (with the point null hypothesis as the denominator).}
\usage{
dir_tests(obj, type = "odds")
}
\arguments{
\item{obj}{A "bma" object.}

\item{type}{"odds" or "probs"}
}
\description{
Compute Bayes factors, posterior odds etc. (or posterior/prior probabilities) for directional hypotheses (with the point null hypothesis as the denominator).
}
\details{
This function is not designed to be used by the analyst and is therefore not exported.
}
