% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc_functions.R
\name{has_needed}
\alias{has_needed}
\title{Test if a set of model terms (fixed effects) contains all the needed terms one order lower than an interaction.}
\usage{
has_needed(terms, intr)
}
\arguments{
\item{terms}{A character vector of model terms.}

\item{intr}{The interaction, written in a form like "a:b" or "x1:x2:x3", etc.}
}
\value{
A logical value indicating whether or not the necessary terms are present.
}
\description{
Test if a set of model terms (fixed effects) contains all the needed terms one order lower than an interaction.
}
