% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_stan_data.R
\name{make_stan_data}
\alias{make_stan_data}
\title{Prepare data for regression by giving factors appropriate contrast codes and (optionally) mean-centering numeric predictors.}
\usage{
make_stan_data(formula, data, prior_scale = 1)
}
\arguments{
\item{formula}{An object of class formula or brmsformula (or one that can be coerced to that classes): A symbolic description of the model to be fitted. The details of model specification are explained in brmsformula.}

\item{data}{A data frame.}
}
\value{
A list with the following elements:
data: a modified data frame that is ready for regression
x_numeric_names: names of data variables that are numeric
x_factor_names: names of data variables that are factors
}
\description{
Prepare data for regression by giving factors appropriate contrast codes and (optionally) mean-centering numeric predictors.
}
\details{
* DESCRIBE THE CONTRAST CODES ETC
}
