% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_formula.R
\name{parse_formula}
\alias{parse_formula}
\title{Parse a regression model formula into its component parts.}
\usage{
parse_formula(formula)
}
\description{
Parse a regression model formula into its component parts.
}
