% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods_and_utilities.R
\name{plot.bma}
\alias{plot.bma}
\title{Plot the posterior distribution of a regression coefficient.}
\usage{
\method{plot}{bma}(obj, coef)
}
\arguments{
\item{obj}{A "bma" object.}

\item{coef}{The coefficient whose posterior distribution you want to plot.}
}
\description{
Plot the posterior distribution of a regression coefficient.
}
\details{
The posterior distribution is approximated with a mixture of normals approximation. Only models that include the coefficient are included, i.e. the distribution plotted is the posterior distribution for the coefficient given that it is non-zero.
}
