% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods.R
\name{plot.breg}
\alias{plot.breg}
\title{Create various types of plot.}
\usage{
\method{plot}{breg}(obj, type = "post_pred", pars = obj$coef_names)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}

\item{type}{The type of plot. See "details" for a list of available options.}

\item{pars}{Parameters to select. By default these are just the model coefficients (fixed effects).}
}
\description{
Create various types of plot.
}
\details{
The following plot types are available:
\itemize{
 \item "post_pred": posterior predictive plot
 \item "intervals": posterior means and credible intervals
 \item "density": posterior distributions
 \item "trace": trace plot for visually assessing MCMC convergence (should look like a "fuzzy caterpillar")
}
Several of these are based on some of the built-in Rstan plotting functions (https://mc-stan.org/rstan/reference/stan_plot.html).
You could use those Rstan plotting functions directly on obj$stanfit for more options and flexibility.
Also, note that the results are ggplot2 objects, so could use ggplot2 to modify them (e.g. add a different title).
}
