% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods.R
\name{posterior_linpred.breg}
\alias{posterior_linpred.breg}
\title{Draw from the posterior distribution of the linear predictor (mu = b0 + b1*x + ...).}
\usage{
\method{posterior_linpred}{breg}(
  obj,
  newdata = NULL,
  ndraws = NULL,
  seed = sample.int(.Machine$integer.max, size = 1L)
)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}

\item{newdata}{Optionally, a data frame in which to look for variables with which to predict. If omitted, the fitted data are used.}

\item{ndraws}{Number of posterior draws to use, i.e. number of replicated data sets to simulate. Defaults to NULL (use all).}

\item{seed}{The seed for random number generation. Set this manually if you want reproducible results.}
}
\value{
A D x N matrix of samples from the posterior distribution of the linear predictor (mu), where D is the number of draws and N is the number of data points.
}
\description{
Draw from the posterior distribution of the linear predictor (mu = b0 + b1*x + ...).
}
