% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.R
\name{prepare_data}
\alias{prepare_data}
\title{Prepare data for being converted to a design matrix: convert character variables to factors, give appropriate contrasts for Bayesian inference, and optionally mean-center numeric predictors.}
\usage{
prepare_data(formula_info, data, center = TRUE)
}
\description{
Prepare data for being converted to a design matrix: convert character variables to factors, give appropriate contrasts for Bayesian inference, and optionally mean-center numeric predictors.
}
