% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probs_term_to_model.R
\name{probs_term_to_model}
\alias{probs_term_to_model}
\title{Convert term inclusion probabilities into model probablities.}
\usage{
probs_term_to_model(main_probs, intr_condprobs, model_info)
}
\arguments{
\item{main_probs}{A named vector of inclusion probabilities for main effects.}

\item{intr_condprobs}{A named vector of inclusion conditional probabilities for interactions, or NULL if there are no interactions.}

\item{model_info}{Information about submodels, returned by using the function "submodels" on "formula(full_model)".}
}
\value{
A named vector of model probabilities (names = model formulas).
}
\description{
Convert term inclusion probabilities into model probablities.
}
\details{
Prior probabilities for main effects are assumed to be independent of each other. Prior probabilities for two-way interactions are conditional on the inclusion of both relevant main effects, those for three-way interactions on the inclusion of relevant main effects and two-way interactions, etc.
}
