% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/restricted_model_bfs_sdr.R
\name{restricted_model_bfs_sdr}
\alias{restricted_model_bfs_sdr}
\title{Use the Savage-Dickey Ratio with a brms model to compute Bayes factors for restricted models.}
\usage{
restricted_model_bfs_sdr(brmsfit)
}
\arguments{
\item{brmsfit}{A fitted brms model.}
}
\value{
A table (data frame) of approximate Bayes factors.
}
\description{
Use the Savage-Dickey Ratio with a brms model to compute Bayes factors for restricted models.
}
\details{
ADD THESE
}
