% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submodels.R
\name{submodels}
\alias{submodels}
\title{Find all the submodels (full model plus restricted models) given a regression model formula.}
\usage{
submodels(formula)
}
\arguments{
\item{formula}{An object of class formula or brmsformula (or one that can be coerced to that classes): A symbolic description of the full model, i.e. the model including all terms being considered. This could also be a string representing the formula.}
}
\value{
A list containing the following elements:
model_names = model formulas as strings
formulas = model formulas
included = terms included in each model
omitted = terms omitted in each model
included_table = same info as "included" except in a data frame
n_models = number of models
n_terms = number of fixed effects
term_names = names of fixed effects. The models are listed in each component of the output in the same order, with the full model given first.
}
\description{
Find all the submodels (full model plus restricted models) given a regression model formula.
}
\details{
This function ignores random effects; appropriate random effects for submodels can be obtained using the "add_random_effects" function. GIVE FURTHER DETAILS**
}
