% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods.R
\name{summary.bma}
\alias{summary.bma}
\title{Summarize information about a "bma" object.}
\usage{
\method{summary}{bma}(obj, type = "terms", pretty = TRUE)
}
\arguments{
\item{obj}{A "bma" object.}

\item{type}{Type of summary. Options are "terms", "est", and "models" (defaults to "terms").}

\item{pretty}{Logical. If TRUE, then the output is printed in an easy to read format, but of the "character" data type. If FALSE then the raw numeric output is returned, without rounding etc.}
}
\description{
Summarize information about a "bma" object.
}
\details{
GIVE INFO ABOUT SUMMARY TYPES
The first model in the list is used as the denominator for model comparison Bayes factors, i.e. the Bayes factor for model i is defined as p(D | first model)/p(D | model i).
}
