% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bma_methods.R
\name{summary.bma}
\alias{summary.bma}
\title{Summarize information about a "bma" object.}
\usage{
\method{summary}{bma}(obj, type = "terms", digits_to_round = 3)
}
\arguments{
\item{obj}{A "bma" object.}

\item{type}{Type of summary. Options are "terms", "est", and "models" (defaults to "terms").}

\item{digits_to_round}{Number of digits to round results to.}
}
\description{
Summarize information about a "bma" object.
}
\details{
GIVE INFO ABOUT SUMMARY TYPES
The first model in the list is used as the denominator for model comparison Bayes factors, i.e. the Bayes factor for model i is defined as p(D | first model)/p(D | model i).
}
