% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods_and_utilities.R
\name{summary.breg}
\alias{summary.breg}
\alias{summary}
\title{Show summary information about a "breg" object.}
\usage{
\method{summary}{breg}(obj)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}
}
\description{
Show summary information about a "breg" object.
}
