% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg_bma_methods.R
\name{summary.reg_bma}
\alias{summary.reg_bma}
\title{Summarize information about a "reg_bma" object.}
\usage{
\method{summary}{reg_bma}(obj, type = "terms", digits_to_round = 3)
}
\arguments{
\item{obj}{A "reg_bma" object.}

\item{type}{Type of summary. Options are "terms", "est", and "models" (defaults to "terms").}

\item{digits_to_round}{Number of digits to round results to.}
}
\description{
Summarize information about a "reg_bma" object.
}
