% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/breg_methods.R
\name{vcov.breg}
\alias{vcov.breg}
\title{Computes the posterior variance-covariance matrix of model parameters.}
\usage{
\method{vcov}{breg}(obj, pars = obj$coef_names)
}
\arguments{
\item{obj}{A "breg" object (fitted model).}

\item{pars}{Parameters to select. By default these are just the model coefficients (fixed effects).}
}
\value{
The posterior variance-covariance matrix.
}
\description{
Computes the posterior variance-covariance matrix of model parameters.
}
